[MISC] Split HA tests (#880)
* Split HA tests


* Fix spread configurations


---------
