+++ tests/spread/test_self_healing_1.py/task.yaml
@@ -1,0 +1,7 @@
+summary: test_self_healing_1.py
+environment:
+  TEST_MODULE: ha_tests/test_self_healing_1.py
+execute: |
+  tox run -e integration -- "tests/integration/$TEST_MODULE" --model testing --alluredir="$SPREAD_TASK/allure-results"
+artifacts:
+  - allure-results
