+++ tests/spread/test_self_healing_3.py/task.yaml
@@ -1,6 +1,6 @@
-summary: test_self_healing.py
+summary: test_self_healing_3.py
 environment:
-  TEST_MODULE: ha_tests/test_self_healing.py
+  TEST_MODULE: ha_tests/test_self_healing_3.py
 execute: |
   tox run -e integration -- "tests/integration/$TEST_MODULE" --model testing --alluredir="$SPREAD_TASK/allure-results"
 artifacts:
